chore: rename plugin.ts to commands.ts
